move converters out of configprovider
